requirements.txt
attrs~=21.2.0
bitarray~=2.3.4
numpy~=1.21.4
pylint~=2.12.2
